requirements.txt
asposestorage==1.0.2
einops==0.4.1
h5py==2.10.0
ipdb==0.13.9
kmeans_pytorch==0.3
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.2
nystrom_attention==0.0.11
opencv_python==4.1.1.26
openslide_python==1.1.2
pandas==1.1.5
Pillow==9.2.0
preprocessing==0.1.13
requests==2.27.1
scikit_learn==1.1.2
scikit_multilearn==0.2.0
scipy==1.5.2
seaborn==0.11.2
skimage==0.0
tensorboardX==2.5.1
tensorflow==2.10.0
timm==0.5.4
torch==1.7.1
torchinfo==1.7.0
torchsampler==0.1.2
torchsummary==1.5.1
torchvision==0.8.2
tqdm==4.64.1
visdom==0.1.8.9
webdataset==0.2.20
